add scan.thread_count configuration value

+++ crates/mds-util/src/resource_scaling.rs
@@ -103,11 +103,11 @@ impl HostResources {
         curr_max_fd
     }
 
-    pub fn max_threads(&mut self) -> usize {
+    pub fn max_threads(&mut self) -> u16 {
         let calculated_threads = self.performance_tier().max_threads();
         let max_threads = calculated_threads.clamp(MIN_LOW_TIER_THREADS, MAX_IO_THREADS);
         let scaled_max_fd = (self.max_file_descriptors() / 3) as usize;
-        max_threads.min(scaled_max_fd)
+        max_threads.min(scaled_max_fd) as u16
     }
 
     pub fn passive_refresh_interval(&mut self) -> Duration {
